requirements.txt
flask
gunicorn
flask-cors
requests
google-auth
diskcache
pydub
google-genai
dotenv
ollama
coqui-tts
groq
streamlit
FastAPI
audio-recorder-streamlit
edge-tts
reportlab
langdetect
deepgram-sdk 